refactor: Move some common models
